// frontend/libs/data-domain/video-data/src/lib/enums/video-form-model-fields.enum.ts
export enum VideoFormModelFieldsEnum {
  NAME = 'name',
  VIDEO_LINK = 'videoLink',
  CHANNEL_ID = 'channelId',
  CATEGORY = 'category',
  UPLOAD_DATE = 'uploadDate',
  DATE_OF_RECORDING = 'dateOfRecording',
  TOPIC = 'topic',
  GUESTS = 'guests',
  WEAPON_TYPE = 'weaponType',
  GAME_SYSTEM = 'gameSystem',
  TOURNAMENT_ID = 'tournamentId',
  TEAM_ONE_ID = 'teamOneId',
  TEAM_TWO_ID = 'teamTwoId',
  TEAM_ONE_CITY = 'teamOneCity',
  TEAM_TWO_CITY = 'teamTwoCity',
  TEAM_ONE_MIX = 'teamOneMix',
  TEAM_TWO_MIX = 'teamTwoMix',
  COMMENT = 'comment',
  TOURNAMENT_CITY = 'tournamentCity',
  TOURNAMENT_START_DATE = 'tournamentStartDate',
  TOURNAMENT_END_DATE = 'tournamentEndDate',
  CHANNEL_LINK = 'channelLink',
};